Extract the `Scope` class
